logout

+++ src/components/LPheader.jsx
@@ -200,10 +200,13 @@ const LPheader = () => {
 
 //================================================================sign out
   const handleLogout = () => {
-    localStorage.clear();
+    if (localStorage.length === 0) {
       window.location.href = "https://core.leadplaner.com/login";
+    } else {
+      localStorage.clear();
+      window.location.href = "https://core.leadplaner.com/login";
+    }
   };
- 
   return (
     <>
       <div className="nav">
